+++ lib/s_spreadsheet/s_spreadsheet.dart
@@ -5,3 +5,6 @@
 library;
 
 export 'src/s_spreadsheet.dart';
+
+// Re-export IndexedScrollController so callers don't need an extra import.
+export 'package:s_packages/indexscroll_listview_builder/indexscroll_listview_builder.dart' show IndexedScrollController;